backend/requirements.txt
langchain-groq
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
pypdf==3.17.0
streamlit==1.28.0
reportlab==4.0.8
python-dotenv==1.0.0
huggingface-hub
fastapi 
uvicorn
python-multipart
sentence-transformers
langchain_huggingface